backend/pyproject.toml
[project]
name = "travel-map-backend"
version = "1.0.0"
description = "Backend for travel_map — print-ready stylized map generator"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "pillow>=11,<12",
    "pycairo>=1.27,<2",
    "numpy>=2,<3",
    "httpx>=0.28,<1",
    "aiosqlite>=0.21,<1",
    "pydantic-settings>=2,<3",
    "python-multipart>=0.0.18,<1",
    "pyyaml>=6,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.25,<1",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
